README.md: Add cluster-machine-approver details.
While working on integration of a new IPI platform type, I needed to
understand how automatic CSR approval worked, because it imposes some
expectations on each platform's cluster-api provider (or Machine
actuator).  This patch is adapted from some of my notes.

This expansion of the README aims to be a document that provides
background and pointers to important details for anyone looking
understand the CSR approval workflows.