Fall back to machine-api check if certificate check fails
This falls back to the original machine-api based checks for serving
certificate renewals if the checks based on retrieving the current
serving certificate fail.  This is needed in cases where network
interfaces are added to a machine after its initial creation, because
the SAN list in the new CSR will contain new addresses and no longer
match the current certificate exactly.

This requires the machine-controller to track all addresses on a
machine, which was recently fixed for AWS:

  https://github.com/openshift/cluster-api-provider-aws/pull/277