CO status controller: Add clusterOperator manifest
The payload build process replace "0.0.1-snapshot" in the manifest with the targeted version.
This manifests is then used by the CVO to compare targeted version against the one in the living clusterOperator object.
The controller is responsible for setting the targeted one in the living clusterOperator object during an upgrade.

https://github.com/openshift/cluster-version-operator/blob/ed67bf413cc1765f36c31da22a72950bf9f3b7eb/docs/dev/clusteroperator.md#how-should-i-include-clusteroperator-custom-resource-in-manifests