Use a dedicated flagset for CMA args
One of the dependencies registers a kubeconfig flag at init. This 
conflicts with the flag we are registering and causes a panic.

If we use our own flagset we can avoid dependencies registering their 
own flags and handle that explicitly ourselves.

This change does not affect any of the existing flag behaviour as all 
flags are carried into the new flagset.