[config] Add disabling node server cert approvals
This commit adds support to disable node server cert auto approvals.

One benefit of this is that the machine approver controller can be stopped from
automatically approving a node's server CSR when it is not the one who approved
the node's client CSR (due to disabled workflow), preventing races with any
other entities that are handling CSRs (e.g. Windows Machine Config Operator).