Patch CSR name onto Machine annotations
Sometimes it may happen that it takes Machine controller some time to
reconcile the Machine and update object with current state while cloud
provider already did so. In such situations automatic approval of the
CSR may fail because Machine object Status is not up to date.

In order to force reconciliation of the Machine object this commit
implements adding an annotation with the CSR name to the Machine object
when CSR verification fails.